Revert moving CSS modules to private
