+++ projects/js-packages/charts/src/components/bar-chart/bar-chart.module.scss
