+++ projects/js-packages/charts/src/components/bar-chart/private/index.ts
@@ -1,2 +1,1 @@
 export { useBarChartOptions } from './use-bar-chart-options';
-export { default as styles } from './bar-chart.module.scss';
