+++ projects/js-packages/charts/src/components/conversion-funnel-chart/conversion-funnel-chart.module.scss
