+++ projects/js-packages/charts/src/components/leaderboard-chart/leaderboard-chart.module.scss
