+++ projects/js-packages/charts/src/components/legend/legend.module.scss
