+++ projects/js-packages/charts/src/components/legend/private/index.ts
@@ -1,2 +1,1 @@
-export { default as styles } from './legend.module.scss';
 export { BaseLegend } from './base-legend';
