+++ projects/js-packages/charts/src/components/line-chart/line-chart.module.scss
