+++ projects/js-packages/charts/src/components/line-chart/line-chart.tsx
@@ -25,7 +25,8 @@ import { DefaultGlyph } from '../private/default-glyph';
 import { SingleChartContext, type SingleChartRef } from '../private/single-chart-context';
 import { withResponsive } from '../private/with-responsive';
 import { AccessibleTooltip, useKeyboardNavigation } from '../tooltip';
-import { LineChartAnnotation, LineChartAnnotationsOverlay, styles } from './private';
+import styles from './line-chart.module.scss';
+import { LineChartAnnotation, LineChartAnnotationsOverlay } from './private';
 import type { BaseChartProps, DataPoint, DataPointDate, SeriesData, Optional } from '../../types';
 import type { ResponsiveConfig } from '../private/with-responsive';
 import type { TickFormatter } from '@visx/axis';
