+++ projects/js-packages/charts/src/components/line-chart/private/index.ts
@@ -1,4 +1,3 @@
-export { default as styles } from './line-chart.module.scss';
 export { default as LineChartAnnotationLabelWithPopover } from './line-chart-annotation-label-popover';
 export { default as LineChartAnnotationsOverlay } from './line-chart-annotations-overlay';
 export { default as LineChartAnnotation } from './line-chart-annotation';
