+++ projects/js-packages/charts/src/components/line-chart/private/line-chart-annotations-overlay.tsx
@@ -1,7 +1,7 @@
 import { DataContext } from '@visx/xychart';
 import { useEffect, useState, useCallback } from 'react';
 import { useSingleChartContext } from '../../private/single-chart-context';
-import styles from './line-chart.module.scss';
+import styles from '../line-chart.module.scss';
 import type { AxisScale } from '@visx/axis';
 import type { FC, ReactNode } from 'react';
