+++ projects/js-packages/charts/src/components/pie-chart/pie-chart.module.scss
