--- projects/js-packages/charts/src/components/pie-chart/private/index.ts
@@ -1,1 +1,0 @@
-export { default as styles } from './pie-chart.module.scss';
