+++ projects/js-packages/charts/src/components/pie-semi-circle-chart/pie-semi-circle-chart.module.scss
