--- projects/js-packages/charts/src/components/pie-semi-circle-chart/private/index.ts
@@ -1,1 +1,0 @@
-export { default as styles } from './pie-semi-circle-chart.module.scss';
