+++ projects/js-packages/charts/src/components/tooltip/base-tooltip.module.scss
