--- projects/js-packages/charts/src/components/tooltip/private/index.ts
@@ -1,1 +1,0 @@
-export { default as styles } from './base-tooltip.module.scss';
